[feat] refactoring and rename myprofilecard
refactoring css with tailwind css
