+++ src/molecules/UserProfileCard/UserProfileCard.module.css
@@ -1,0 +1,45 @@
+@import "src/molecules/ProfileCard/ProfileCard.module";
+
+.my_profile_card {
+  @apply flex justify-between items-center relative;
+  @apply bg-white;
+
+  /* 반응형 */
+  /* 기본: 모바일 (~765) */
+  @apply min-h-[5.5rem] p-4 gap-4;
+  /* md: 태블릿 (768+) */
+  @apply md:min-h-[5.5rem] md:p-4 md:gap-4;
+  /* lg: 데스크탑 (1024+) */
+  @apply lg:min-h-[6rem] lg:p-4 lg:gap-4;
+  /* 2xl: 데스크탑 (1536+) */
+  @apply 2xl:min-w-[6rem] 2xl:p-6 2xl:gap-6;
+}
+/* TODO: 좁아지면 겹치는거 해결해야함 */
+
+.content {
+  @apply flex items-center;
+
+  /* 반응형 */
+  /* 기본: 모바일 (~765) */
+  @apply gap-3;
+  /* md: 태블릿 (768+) */
+  @apply md:gap-3;
+  /* lg: 데스크탑 (1024+) */
+  @apply lg:gap-4;
+  /* 2xl: 데스크탑 (1536+) */
+  @apply 2xl:gap-5;
+}
+
+.icons {
+  @apply justify-between items-center gap-2 absolute;
+
+  /* 반응형 */
+  /* 기본: 모바일 (~765) */
+  @apply hidden;
+  /* md: 태블릿 (768+) */
+  @apply md:hidden;
+  /* lg: 데스크탑 (1024+) */
+  @apply lg:flex lg:right-4 lg:bottom-1;
+  /* 2xl: 데스크탑 (1536+) */
+  @apply 2xl:flex 2xl:right-6 2xl:bottom-3;
+}
